+++ CHANGELOG.md
@@ -8,8 +8,9 @@
   - Resolve full screen menubar covering the app toolbar
   - Resolve unfocused password windows
 - Resolve USB 1.1 kernel panics on macOS 14.1
+- Resolve PCIe FaceTime camera support on macOS 14.1
 - Increment Binaries:
-  - PatcherSupportPkg 1.3.6 - release
+  - PatcherSupportPkg 1.3.7 - release
   - AirportBrcmFixup 2.1.8 - release
   - BlueToolFixup 2.6.8 - release
   - RestrictEvents 1.1.3 - release